refactor: restructure application by replacing index.js with app.js as main entry point and updating related configurations fix: modify image field in article model to be optional fix: correct multer storage path for image uploads fix: update error handling in middleware to return proper status codes add: implement utility function for deleting images from the filesystem add: enhance article routes with new endpoints for fetching articles by criteria fix: populate comments with associated articles in getComments controller fix: validate ObjectId in user and city controllers

+++ controllers/article.controller.js
@@ -1,11 +1,8 @@
-import Article from "../models/article.model.js";
-import path from "path";
-import fs from "fs";
 import mongoose from "mongoose";
+import Article from "../models/article.model.js";
+import { deleteImage } from "../utiles/utlies.js";
 
 const createArticle = async (req, res, next) => {
-  const image = req.file.filename;
-
   try {
     const {
       title,
@@ -15,13 +12,17 @@ const createArticle = async (req, res, next) => {
       userId,
       subArticles,
       phone,
+      whatsUp,
     } = req.body;
 
     let parsedSubArticles = Array.isArray(subArticles)
       ? subArticles
       : JSON.parse(subArticles || "[]");
 
-    const mainImage = req.files["image"]?.[0];
+    const mainImage = req.files["image"]?.[0].filename
+      ? `/images/${req.files["image"]?.[0].filename}`
+      : null;
+
     const subArticleImages = req.files["subArticleImages"] || [];
 
     if (!mainImage) {
@@ -29,26 +30,37 @@ const createArticle = async (req, res, next) => {
       error.statusCode = 400;
       throw error;
     }
-
-    const subArticlesWithImages = parsedSubArticles.map((subArticle, index) => {
-      const imageFile = subArticleImages[index];
+    let imageIndex = 0;
+    const subArticlesWithImages = parsedSubArticles.map((subArticle) => {
+      if (subArticle.image && imageIndex < subArticleImages.length) {
+        const imageFile = subArticleImages[imageIndex];
+        imageIndex++;
+        return {
+          ...subArticle,
+          image: `/images/${imageFile.filename}`,
+        };
+      }
       return {
         ...subArticle,
-        image: imageFile.path || "",
+        image: subArticle.image || "",
       };
     });
+
+    console.log(subArticlesWithImages, "subArticlesWithImages");
 
     const article = await Article.create({
       title,
       description,
-      image,
+      image: mainImage,
       phone,
       user: userId,
       service: serviceId,
       city: cityId,
       subArticles: subArticlesWithImages,
+      whatsUp,
     });
-    res.status(2001).json({
+
+    res.status(201).json({
       success: true,
       data: article,
     });
@@ -59,10 +71,124 @@ const createArticle = async (req, res, next) => {
 
 const getArticles = async (req, res, next) => {
   try {
-    const articles = await Article.find();
-    res.status(200).json({
+    const limit = parseInt(req.query.limit) || null;
+    const page = parseInt(req.query.page) || null;
+    const skip = (page - 1) * limit;
+
+    const total = await Article.countDocuments();
+
+    if (limit && page) {
+      const articles = await Article.find()
+        .skip(skip)
+        .limit(limit)
+        .populate("service")
+        .populate("city")
+        .populate("user", "name email");
+      return res.status(200).json({
+        success: true,
+        data: articles,
+        pagination: {
+          current_page: page,
+          per_page: limit,
+          total,
+          total_pages: Math.ceil(total / limit),
+        },
+      });
+    }
+
+    const articles = await Article.find()
+      .populate("service")
+      .populate("city")
+      .populate("user", "name email");
+
+    return res.status(200).json({
       success: true,
       data: articles,
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const searchArticles = async (req, res, next) => {
+  try {
+    const limit = parseInt(req.query.limit) || 10;
+    const page = parseInt(req.query.page) || 1;
+    const skip = (page - 1) * limit;
+    const text = req.query.q;
+
+    const total = await Article.countDocuments();
+
+    const articles = await Article.aggregate([
+      {
+        $lookup: {
+          from: "services",
+          localField: "service",
+          foreignField: "_id",
+          as: "service",
+        },
+      },
+      { $unwind: { path: "$service", preserveNullAndEmptyArrays: true } },
+      {
+        $lookup: {
+          from: "cities",
+          localField: "city",
+          foreignField: "_id",
+          as: "city",
+        },
+      },
+      { $unwind: { path: "$city", preserveNullAndEmptyArrays: true } },
+      {
+        $lookup: {
+          from: "users",
+          localField: "user",
+          foreignField: "_id",
+          as: "user",
+        },
+      },
+      { $unwind: { path: "$user", preserveNullAndEmptyArrays: true } },
+      {
+        $match: {
+          $or: [
+            { title: { $regex: text, $options: "i" } },
+            { description: { $regex: text, $options: "i" } },
+            { "subArticles.title": { $regex: text, $options: "i" } },
+            { "subArticles.description": { $regex: text, $options: "i" } },
+            { "service.name": { $regex: text, $options: "i" } },
+            { "city.name": { $regex: text, $options: "i" } },
+          ],
+        },
+      },
+      {
+        $project: {
+          title: 1,
+          description: 1,
+          subArticles: 1,
+          views: 1,
+          phone: 1,
+          whatsUp: 1,
+          image: 1,
+          service: 1,
+          city: 1,
+          "user.name": 1,
+          "user.email": 1,
+          createdAt: 1,
+          updatedAt: 1,
+        },
+      },
+      { $skip: skip },
+      { $limit: parseInt(limit) },
+    ]);
+
+    return res.status(200).json({
+      success: true,
+      data: articles,
+      pagination: {
+        current_page: page,
+        per_page: limit,
+        total,
+        total_pages: Math.ceil(total / limit),
+      },
     });
   } catch (error) {
     next(error);
@@ -89,16 +215,90 @@ const getArticle = async (req, res, next) => {
       {
         new: true,
       }
-    );
+    )
+      .populate("service")
+      .populate("city")
+      .populate("user", "name email");
+
     if (!article) {
-      const error = new Error("City not found");
+      const error = new Error("service not found");
       error.statusCode = 404;
       throw error;
     }
+    const relatedArticles = await Article.find({
+      service: article.service._id,
+    })
+      .limit(3)
+      .populate("service")
+      .populate("city")
+      .populate("user", "name email");
+
     res.status(200).json({
       success: true,
       data: article,
+      related: relatedArticles,
     });
+  } catch (error) {
+    next(error);
+  }
+};
+const getArticleBy = async (req, res, next) => {
+  try {
+    const { cityId, serviceId } = req.body;
+    const limit = parseInt(req.query.limit) || 10;
+    const page = parseInt(req.query.page) || 1;
+    const skip = (page - 1) * limit;
+    if (serviceId && !mongoose.Types.ObjectId.isValid(serviceId)) {
+      const error = new Error(" invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
+
+    if (cityId && !mongoose.Types.ObjectId.isValid(cityId)) {
+      const error = new Error(" invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
+
+    if (serviceId) {
+      const article = await Article.find({
+        service: serviceId,
+      })
+        .skip(skip)
+        .limit(limit)
+        .populate("service")
+        .populate("city")
+        .populate("user", "name email");
+
+      if (!article) {
+        const error = new Error("City not found");
+        error.statusCode = 404;
+        throw error;
+      }
+      res.status(200).json({
+        success: true,
+        data: article,
+      });
+    }
+    if (cityId) {
+      const article = await Article.find({
+        city: cityId,
+      })
+        .skip(skip)
+        .limit(limit)
+        .populate("service")
+        .populate("city")
+        .populate("user", "name email");
+      if (!article) {
+        const error = new Error("City not found");
+        error.statusCode = 404;
+        throw error;
+      }
+      res.status(200).json({
+        success: true,
+        data: article,
+      });
+    }
   } catch (error) {
     next(error);
   }
@@ -107,6 +307,11 @@ const getArticle = async (req, res, next) => {
 const deleteArticle = async (req, res, next) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
     const article = await Article.findByIdAndDelete(id);
     if (!article) {
       const error = new Error("City not found");
@@ -136,7 +341,17 @@ const updateArticle = async (req, res, next) => {
       whatsUp,
     } = req.body;
 
-    const article = await Article.findById(id);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
+
+    const article = await Article.findById(id)
+      .populate("service")
+      .populate("city")
+      .populate("user", "name email");
+
     if (!article) {
       const error = new Error(`Article not found`);
       error.statusCode = 404;
@@ -150,23 +365,29 @@ const updateArticle = async (req, res, next) => {
     let oldMainImagePath = article.image;
     const oldSubArticleImagePaths = article.subArticles.map((sub) => sub.image);
 
-    const mainImage = req.files["image"]?.[0];
+    const mainImage = req.files["image"]?.[0].filename
+      ? `/images/${req.files["image"]?.[0].filename}`
+      : undefined;
+
     const subArticleImages = req.files["subArticleImages"] || [];
 
-    if (!mainImage) {
-      const error = new Error("article image not found");
-      error.statusCode = 404;
-      throw error;
-    }
-
-    const subArticlesWithImages = parsedSubArticles.map((subArticle, index) => {
-      const imageFile = subArticleImages[index];
-      if (!imageFile) {
-        throw new Error(`Image missing for subArticle at index ${index}`);
+    let imageIndex = 0;
+    const subArticlesWithImages = parsedSubArticles.map((subArticle) => {
+      if (
+        subArticle.image &&
+        !subArticle.image.startsWith("/images") &&
+        imageIndex < subArticleImages.length
+      ) {
+        const imageFile = subArticleImages[imageIndex];
+        imageIndex++;
+        return {
+          ...subArticle,
+          image: `/images/${imageFile.filename}`,
+        };
       }
       return {
         ...subArticle,
-        image: imageFile.path,
+        image: subArticle.image || "",
       };
     });
 
@@ -178,20 +399,20 @@ const updateArticle = async (req, res, next) => {
     article.city = cityId || article.city;
     article.user = userId || article.user;
     article.subArticles = subArticlesWithImages;
-    article.image = mainImage.path || article.image;
+    article.image = mainImage || article.image;
     await article.save();
-    if (mainImage && oldMainImagePath !== article.image) {
-      const fullOldPath = path.join(__dirname, oldMainImagePath);
-      if (fs.existsSync(fullOldPath)) {
-        await fs.unlink(fullOldPath);
-      }
+
+    if (mainImage) {
+      await deleteImage(oldMainImagePath);
     }
 
     await Promise.all(
-      oldSubArticleImagePaths.map(async (imagePath) => {
-        const fullPath = path.join(__dirname, imagePath);
-        if (fs.existsSync(fullPath)) {
-          await fs.unlink(fullPath);
+      oldSubArticleImagePaths.map(async (oldImagePath) => {
+        const isStillUsed = subArticlesWithImages.some(
+          (sub) => sub.image === oldImagePath
+        );
+        if (!isStillUsed) {
+          await deleteImage(oldImagePath);
         }
       })
     );
@@ -211,4 +432,80 @@ const updateArticle = async (req, res, next) => {
   }
 };
 
-export { getArticles, getArticle, deleteArticle, createArticle, updateArticle };
+const getLimitedArticles = async (req, res, next) => {
+  try {
+    const articles = await Article.aggregate([
+      {
+        $lookup: {
+          from: "services",
+          localField: "service",
+          foreignField: "_id",
+          as: "service",
+        },
+      },
+      { $unwind: "$service" },
+      {
+        $lookup: {
+          from: "cities",
+          localField: "city",
+          foreignField: "_id",
+          as: "city",
+        },
+      },
+      { $unwind: "$city" },
+      {
+        $lookup: {
+          from: "users",
+          localField: "user",
+          foreignField: "_id",
+          as: "user",
+        },
+      },
+      { $unwind: "$user" },
+      {
+        $project: {
+          "user.name": 1,
+          "user.email": 1,
+          service: 1,
+          city: 1,
+          title: 1,
+          description: 1,
+          views: 1,
+          image: 1,
+          createdAt: 1,
+        },
+      },
+      {
+        $group: {
+          _id: "$service._id",
+          service: { $first: "$service" },
+          articles: { $push: "$$ROOT" },
+        },
+      },
+      {
+        $project: {
+          service: 1,
+          articles: { $slice: ["$articles", 10] },
+        },
+      },
+    ]);
+
+    return res.status(200).json({
+      success: true,
+      data: articles,
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+export {
+  createArticle,
+  deleteArticle,
+  getArticle,
+  getArticles,
+  searchArticles,
+  updateArticle,
+  getArticleBy,
+  getLimitedArticles,
+};
