+++ controllers/city.controller.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
 import City from "../models/city.model.js";
+import Article from "../models/article.model.js";
 
 const createCity = async (req, res, next) => {
   const session = await mongoose.startSession();
@@ -47,6 +48,16 @@ const getCities = async (req, res, next) => {
 const getCity = async (req, res, next) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
+
+    const limit = parseInt(req.query.limit) || 10;
+    const page = parseInt(req.query.page) || 1;
+    const skip = (page - 1) * limit;
+
     const existCity = await City.findById(id);
     if (!existCity) {
       const error = new Error("City not found");
@@ -54,9 +65,26 @@ const getCity = async (req, res, next) => {
       throw error;
     }
 
+    const total = await Article.countDocuments({ city: existCity._id });
+
+    const cityArticles = await Article.find({
+      city: existCity._id,
+    })
+      .skip(skip)
+      .limit(limit);
+
     res.status(200).json({
       success: true,
-      data: existCity,
+      data: {
+        city: existCity,
+        articles: cityArticles,
+        pagination: {
+          current_page: page,
+          per_page: limit,
+          total,
+          total_pages: Math.ceil(total / limit),
+        },
+      },
     });
   } catch (error) {
     next(error);
@@ -66,6 +94,11 @@ const getCity = async (req, res, next) => {
 const updateCity = async (req, res, next) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
     const { name } = req.body;
     const updatedCity = await City.findByIdAndUpdate(
       id,
@@ -97,6 +130,11 @@ const updateCity = async (req, res, next) => {
 const deleteCity = async (req, res, next) => {
   try {
     const id = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("invalid id");
+      error.statusCode = 400;
+      throw error;
+    }
     const deletedCity = await City.findByIdAndDelete(id);
 
     if (!deletedCity) {
